requirements.txt
qrcode
flask
gunicorn
python-docx
segno
psycopg2
python-dotenv
dropbox